Changes to be committed: modified:   app/Controllers/Api/AuthController.php modified:   app/Controllers/Api/JobsController.php modified:   app/Database/Migrations/2022-09-02-142915_Users.php modified:   app/Models/Jobs.php

+++ app/Controllers/Api/JobsController.php
@@ -15,6 +15,7 @@ class JobsController extends ResourceController
     public function __construct()
     {
         $this->jobs = new Jobs();
+        $this->users = new Users();
     }
 
     public function index()
@@ -167,15 +168,20 @@ class JobsController extends ResourceController
 			}
 
             $data = $this->jobs->where('job_id', $id)->findAll();
-            if($data){
-            $this->jobs->delete($id);
-                $response = [
-                    'status'   => 200,
-                    'error'    => null,
-                    'messages' => [
-                        'success' => 'Data pekerjaan berhasil dihapus'
-                    ]
-                ];
+            $cek = $this->users->where('job_id', $id)->findAll();
+            if($cek){
+                $this->jobs->updateDataUsers($id);
+
+                if($data){
+                    $this->jobs->delete($id);
+                    $response = [
+                        'status'   => 200,
+                        'error'    => null,
+                        'messages' => [
+                            'success' => 'Data pekerjaan berhasil dihapus'
+                        ]
+                    ];
+                }
             }
             return $this->respondDeleted($response);
         } catch (\Throwable $th) {
